+++ src/collections/aliases/server/publish.js
@@ -2,6 +2,6 @@
 import { rolesAliases } from '../aliases.js';
 
 // returns the list of known contents
-Meteor.publish( 'pwixRoles.Aliases.listAl', function(){
+Meteor.publish( 'Roles.Aliases.listAl', function(){
     return rolesAliases.find();
 });