+++ src/common/i18n/en_US.js
@@ -2,8 +2,8 @@
  * pwix:roles/src/common/i18n/en_US.js
  */
 
-pwixRoles.i18n = {
-    ...pwixRoles.i18n,
+Roles.i18n = {
+    ...Roles.i18n,
     ...{
         en_US: {
             dialogs: {