+++ src/common/js/global.js
@@ -2,4 +2,4 @@
  * pwix:roles/src/common/js/roles.js
  */
 
-pwixRoles = {};
+Roles = {};