+++ src/common/js/i18n.js
@@ -5,15 +5,15 @@
 import { pwixI18n } from 'meteor/pwix:i18n';
 
 import '../i18n/en_US.js';
-pwixI18n.namespace( I18N, 'en', pwixRoles.i18n.en_US );
+pwixI18n.namespace( I18N, 'en', Roles.i18n.en_US );
 
 import '../i18n/fr_FR.js';
-pwixI18n.namespace( I18N, 'fr', pwixRoles.i18n.fr_FR );
+pwixI18n.namespace( I18N, 'fr', Roles.i18n.fr_FR );
 
 /**
  * @locus Anywhere
  * @returns {String} the i18n namespace of this package
  */
-pwixRoles.i18n.namespace = function(){
+Roles.i18n.namespace = function(){
     return I18N;
 }