+++ ChangeLog.md
@@ -6,13 +6,14 @@
 
     Release date: 
 
-    - Define new pwixRoles.countByRole() publication (bumping candidate version number)
+    - Define new Roles.countByRole() publication (bumping candidate version number)
     - Reorganize constants definitions to not pollute global space
     - Bump version requirements Meteor@2.13.2, pwix:i18n@1.5.2, pwix:modal@1.7.1
     - prEdit: fix clickability on disabled nodes
     - prEdit: make sure we edit the account only if a user was provided
     - Remove 'md-modal' mentions from the stylesheet
     - Back to Meteor 2.9.0
+    - pwixRoles global is renamed as Roles
 
 ### 1.1.3
 
@@ -36,7 +37,7 @@
 
     Release date: 2023- 6-20
 
-    - Define pwixRoles.i18n.namespace() which returns the i18n namespace of the package (todo #8)
+    - Define Roles.i18n.namespace() which returns the i18n namespace of the package (todo #8)
     - Replace merge and deep-equal dependencies with lodash
     - configure() now acts both as a getter and a setter